Move utility functions into new file
