+++ src/qibochem/measurement/util.py
@@ -1,0 +1,74 @@
+"""
+Utility functions for optimising measurements and calculation of expectation value
+"""
+
+import networkx as nx
+import numpy as np
+
+
+def check_terms_commutativity(term1: str, term2: str, qubitwise: bool):
+    """
+    Check if terms 1 and 2 are mutually commuting. The 'qubitwise' flag determines if the check is for general
+    commutativity (False), or the stricter qubitwise commutativity.
+
+    Args:
+        term1/term2: Strings representing a single Pauli term. E.g. "X0 Z1 Y3". Obtained from a Qibo SymbolicTerm as
+        ``" ".join(factor.name for factor in term.factors)``.
+        qubitwise (bool): Determines if the check is for general commutativity, or the stricter qubitwise commutativity
+
+    Returns:
+        bool: Do terms 1 and 2 commute?
+    """
+    # Get a list of common qubits for each term
+    common_qubits = {_term[1:] for _term in term1.split() if _term[0] != "I"} & {
+        _term[1:] for _term in term2.split() if _term[0] != "I"
+    }
+    if not common_qubits:
+        return True
+    # Get the single Pauli operators for the common qubits for both Pauli terms
+    term1_ops = [_op for _op in term1.split() if _op[1:] in common_qubits]
+    term2_ops = [_op for _op in term2.split() if _op[1:] in common_qubits]
+    if qubitwise:
+        # Qubitwise: Compare the Pauli terms at the common qubits. Any difference => False
+        return all(_op1 == _op2 for _op1, _op2 in zip(term1_ops, term2_ops))
+    # General commutativity:
+    # Get the number of single Pauli operators that do NOT commute
+    n_noncommuting_ops = sum(_op1 != _op2 for _op1, _op2 in zip(term1_ops, term2_ops))
+    # term1 and term2 have general commutativity iff n_noncommuting_ops is even
+    return n_noncommuting_ops % 2 == 0
+
+
+def group_commuting_terms(terms_list, qubitwise):
+    """
+    Groups the terms in terms_list into as few groups as possible, where all the terms in each group commute
+    mutually == Finding the minimum clique cover (i.e. as few cliques as possible) for the graph whereby each node
+    is a Pauli string, and an edge exists between two nodes iff they commute.
+
+    This is equivalent to the graph colouring problem of the complement graph (i.e. edge between nodes if they DO NOT
+    commute), which this function follows.
+
+    Args:
+        terms_list: List of strings. The strings should follow the output from
+            ``" ".join(factor.name for factor in term.factors)``, where term is a Qibo SymbolicTerm. E.g. "X0 Z1".
+        qubitwise: Determines if the check is for general commutativity, or the stricter qubitwise commutativity
+
+    Returns:
+        list: Containing groups (lists) of Pauli strings that all commute mutually
+    """
+    G = nx.Graph()
+    # Complement graph: Add all the terms as nodes first, then add edges between nodes if they DO NOT commute
+    G.add_nodes_from(terms_list)
+    G.add_edges_from(
+        (term1, term2)
+        for _i1, term1 in enumerate(terms_list)
+        for _i2, term2 in enumerate(terms_list)
+        if _i2 > _i1 and not check_terms_commutativity(term1, term2, qubitwise)
+    )
+    # Solve using Greedy Colouring on NetworkX
+    sorted_groups = nx.coloring.greedy_color(G)
+    group_ids = set(sorted_groups.values())
+    # Sort results so that test results will be replicable
+    term_groups = sorted(
+        sorted(group for group, group_id in sorted_groups.items() if group_id == _id) for _id in group_ids
+    )
+    return term_groups
